utils/db_connection.py
import os
import sqlalchemy
from dotenv import load_dotenv

load_dotenv()

def get_db_engine():
    DB_USER = os.getenv("DB_USER")
    DB_PASSWORD = os.getenv("DB_PASSWORD")
    DB_HOST = os.getenv("DB_HOST")
    DB_NAME = os.getenv("DB_NAME")

    connection_string = f"mysql+pymysql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}/{DB_NAME}"
    engine = sqlalchemy.create_engine(connection_string)
    return engine
