requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
prophet
matplotlib
seaborn
plotly
folium
pydeck
requests
python-dotenv
SQLAlchemy
pymysql
psycopg2
cmdstanpy
category_encoders
joblib
mysql-connector-python
sqlalchemy
prophet